+++ src/main/java/pl/asie/lib/chat/ChatHandler.java
@@ -6,9 +6,10 @@ import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 import net.minecraft.command.CommandBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.util.ChatComponentText;
 import net.minecraft.util.EnumChatFormatting;
+import net.minecraft.util.IChatComponent;
 import net.minecraft.world.WorldServer;
+import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.event.ServerChatEvent;
 import org.apache.logging.log4j.LogManager;
@@ -66,13 +67,13 @@ public class ChatHandler {
 			return;
 		}
 
-		ChatComponentText chat;
+		IChatComponent chat;
 		boolean disableRadius = false;
 		String username = ChatUtils.color(AsieLibMod.nick.getNickname(event.username)) + EnumChatFormatting.RESET;
 		String message = event.message;
 		int dimensionId = event.player.worldObj.provider.dimensionId;
 
-		if(event.message.startsWith("!") && enableShout) {
+		if(enableShout && event.message.startsWith("!")) {
 			message = message.substring(1);
 		}
 
@@ -102,11 +103,11 @@ public class ChatHandler {
 			}
 		}
 
-		if(event.message.startsWith("!") && enableShout) {
-			chat = new ChatComponentText(EnumChatFormatting.YELLOW + shoutPrefix + " " + formattedMessage);
+		if(enableShout && event.message.startsWith("!")) {
+			chat = ForgeHooks.newChatWithLinks(EnumChatFormatting.YELLOW + shoutPrefix + " " + formattedMessage);
 			disableRadius = true;
 		} else {
-			chat = new ChatComponentText(formattedMessage);
+			chat = ForgeHooks.newChatWithLinks(formattedMessage);
 		}
 
 		boolean useRadius = CHAT_RADIUS > 0 && !disableRadius;