+++ changelog.md
@@ -2,6 +2,7 @@
 
   * ADDED: Prattle support
     - Chat tweaks work in every channel now!
+  * ADDED: Config option to make the /realname command require Op rights on a server
   * ADDED: Support for OpenComputers, AppliedEnergistics and Mekanism wrenches
   * CHANGED: Changed how tools from other mods are handled. This might break addons not updated to this version!